added сhips in officeList
