+++ src/components/NavVar/NavBar.tsx
@@ -7,32 +7,22 @@ import ExploreIcon from "@mui/icons-material/Explore";
 import { Stack } from "@mui/material";
 
 import {
-  setUserLocation,
-  setUserLocationWatchId,
+  setDrawerOpen,
 } from "../../redux/UserLocationSlice/UserLocationSlice";
 import { useAppDispatch } from "../../redux/hooks";
+import {DRAWER_TYPES} from '../../types';
 
-const NavBar = () => {
+const NavBar = ({ setMapCenter }: any) => {
   const dispatch = useAppDispatch();
 
-  const handleUserGeoReceive = (position: any) => {
-    dispatch(
-      setUserLocation({
-        lat: position.coords.latitude,
-        lng: position.coords.longitude,
-      })
-    );
+  const handleFilterOpen = () => {
+    dispatch(setDrawerOpen(DRAWER_TYPES.FILTER))
   };
 
-  const handleUserGeoRequest = () => {
-    if ("geolocation" in navigator) {
-      navigator.geolocation.getCurrentPosition(handleUserGeoReceive);
-
-      const userLocationWatchId =
-        navigator.geolocation.watchPosition(handleUserGeoReceive);
-      dispatch(setUserLocationWatchId(userLocationWatchId));
-    } else {
-      alert("Гео недоступно");
+  const handleLocationClick = () => {
+    const map = document.querySelector('#map');
+    if (map) {
+      setMapCenter();
     }
   };
 
@@ -48,7 +38,7 @@ const NavBar = () => {
           direction={"row"}
           width={"100%"}
         >
-          <IconButton>
+          <IconButton onClick={handleFilterOpen}>
             <TuneIcon
               sx={{
                 color: "#000",
@@ -70,7 +60,7 @@ const NavBar = () => {
               }}
             />
           </IconButton>
-          <IconButton onClick={handleUserGeoRequest}>
+          <IconButton onClick={handleLocationClick}>
             <ExploreIcon
               sx={{
                 color: "#000",
