+++ src/components/OfficeList/OfficeItem.css
@@ -1,0 +1,15 @@
+.branch-busy-info .busy-level-red {
+  background-color: #811111;
+  color: #FFFFFF;
+}
+
+.branch-busy-info .busy-level-semi-red {
+  background-color: #f09796;
+  color: #000000;
+}
+
+
+.branch-busy-info .busy-level-light {
+  background-color: #fadede;
+  color: #000000;
+}
