+++ src/consts/index.ts
@@ -5,3 +5,18 @@ export const BusyBranchColor = {
 };
 
 export const busyBranchColors = Object.values(BusyBranchColor);
+
+export const BranchBusyLabel = {
+  Full: 'Высокая',
+  SemiFull: 'Средняя',
+  Empty: 'Низкая'
+};
+
+export const BranchBusyClass = {
+  Full: 'red',
+  SemiFull: 'semi-red',
+  Empty: 'light',
+}
+
+export const busyLabels = Object.values(BranchBusyLabel);
+export const busyClassNames = Object.values(BranchBusyClass);
