+++ src/index.tsx
@@ -9,11 +9,9 @@ const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 root.render(
-  <React.StrictMode>
-    <Provider store={store}>
-        <BrowserRouter>
-          <App />
-        </BrowserRouter>
-    </Provider>
-  </React.StrictMode>
+  <Provider store={store}>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+  </Provider>
 );
